refactor: move load into db_setup
