+++ backend/configs/database.py
@@ -78,15 +78,6 @@ def get_redis_client_generator(
     return lambda: redis.asyncio.Redis.from_url(get_redis_url(db_name))
 
 
-async def load_scripts(db_name: str, scripts_dir: str = "scripts/") -> None:
-    async with get_redis_client_generator(db_name)() as client:
-        for filename in os.listdir(scripts_dir):
-            filepath = os.path.join(scripts_dir, filename)
-            with open(filepath, "r") as f:
-                lua_script_content = f.read()
-            await client.script_load(lua_script_content)
-
-
 def get_scripts(scripts_dir: str = "scripts/") -> dict[str, str]:
     sha1s = {}
     for filename in os.listdir(scripts_dir):
